% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accent.R
\name{accent}
\alias{accent}
\title{Get rid of Accent Marks}
\usage{
accent(x)
}
\arguments{
\item{x}{is a data object which contains diacritical marks.}
}
\value{
a vector with same length of \code{x} without diacritic.
}
\description{
Replace lower and upper case accented letters with their counterpart without diacritical marks.
}
\details{
This function can replace a variety of common marks, but not all of them. It is designed to be expanded on demand.
}
\examples{
x <- 'Thérèse, ne me jugez pas de moitié si vous ne me connaissez pas du tout.'
accent(x)
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}.
}
\seealso{
\code{\link{ascii.table}}.
}
\keyword{Attributes}

