% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{ascii.table}
\alias{ascii.table}
\title{ASCII Characters Table}
\usage{
ascii.table(x)
}
\arguments{
\item{x}{A string whose characters is to be checked against.}
}
\value{
Returns \code{TRUE} if ASCII character and \code{FALSE} otherwise.
}
\description{
To detect ASCII characters, we may need to specify them literally. This function helps identifying what character is in ascii format and what is not.
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

