% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/soundexBR.R
\encoding{UTF-8}
\name{char2int}
\alias{char2int}
\title{Character to Integer}
\usage{
char2int(x)
}
\arguments{
\item{x}{A vector consisting of characters.}
}
\description{
Declares characters to integer.
}
\examples{
char2int("This should be whole numbers")
}
\seealso{
\link{int2char}, \link[base]{iconv}.
}
\keyword{Encoding}

