% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/soundexBR.R
\encoding{UTF-8}
\name{int2char}
\alias{int2char}
\title{Interger to Character}
\usage{
int2char(x)
}
\arguments{
\item{x}{A vector consisting of whole numbers.}
}
\description{
Declares integer inputs as UTF-8 output.
}
\examples{
int2char(c(84,104,105,115,32,115,104,111,117,108,
100, 32, 98, 101, 32, 119, 104, 111, 108, 101,
32, 110, 117, 109, 98, 101, 114, 115))
}
\seealso{
\link{char2int}, \link[base]{iconv}.
}
\keyword{Encoding}

