\encoding{UTF-8}
\name{SoundexBR-package}
\alias{SoundexBR}
\docType{package}
\title{ Phonetic Coding Algorithm For Brazilian Portuguese. 
}
\description{
SoundexBR returns a census-like phonetic code of a string name given the Portuguese sound system. This function was outlined to work with the \code{RecordLinkage} package, however, it is also helpful as a standalone function as it can help on identifying similar names based on a distribution of characters distance.  See details bellow.
}
\details{
\tabular{ll}{
Package: \tab SoundexBR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-12\cr
License: \tab AGPL-3\cr}
}
\author{
Daniel Marcelino <dmarcelino@live.com>
}
\references{
Borg, Andreas and Murat Sariyar. (2012) \emph{RecordLinkage: Record Linkage in R,} R package version 0.4-1,  \url{http://CRAN.R-project.org/package=RecordLinkage}.

Marcelino, Daniel (2013) \emph{SciencesPo: A Tool Set for Analyzing Political Behaviour Data}, \url{(http://dx.doi.org/10.2139/ssrn.2320547}.
}

\keyword{Soundex}