% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{accent}
\alias{accent}
\title{Get rid of Accent Marks}
\usage{
accent(x)
}
\arguments{
\item{x}{is a data object which contains diacritical marks.}
}
\value{
a vector with same length of \code{x} without diacritic.
}
\description{
Replace lower and upper case accented letters with their counterpart without diacritical marks.
}
\details{
This function can replace a variety of common marks, but not all of them. It is designed to be expanded on demand.
}
\examples{
x <- 'Thérèse, não me julgues pela metade se não me conheces por inteiro'
accent(x)
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}.
}
\seealso{
\code{\link{ascii.table}}.
}
\keyword{attributes}

