\name{char2int}
\alias{char2int}
\title{ Character to Integer 
}
\description{ Declares characters to integer.
}
\usage{
char2int(x)
}
\arguments{
  \item{x}{A vector consisting of characters.
}
}

\author{Daniel Marcelino
}

\seealso{\code{\link{int2char}, \code{\link[base]{iconv}}}.
}
\examples{ 
	
	char2int("This should be whole numbers")
	}

\keyword{ Soundex }